% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgeocoder_check_access.R
\name{arc_api_call}
\alias{arc_api_call}
\title{Helper function for centralize API queries}
\usage{
arc_api_call(url, destfile, quiet)
}
\arguments{
\item{url}{a \code{\link{character}} string (or longer vector
    for the \code{"libcurl"} method) naming the URL of a resource to be
    downloaded.}

\item{destfile}{a character string (or vector, see the \code{url}
    argument) with the file path where the downloaded file is to be
    saved.  Tilde-expansion is performed.}

\item{quiet}{If \code{TRUE}, suppress status messages (if any), and
    the progress bar.}
}
\value{
A logical \code{TRUE/FALSE}
}
\description{
A wrapper of \code{\link[utils:download.file]{utils::download.file()}}. On warning on error it will
retry the call.
}
\seealso{
Other api_management: 
\code{\link{arcgeocoder_check_access}()}
}
\concept{api_management}
\keyword{internal}
