% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{arc_categories}
\alias{arc_categories}
\title{Place categories supported by the ArcGIS REST API}
\format{
A \link[tibble:tbl_df]{tibble} with
383 rows and
three variables:
\describe{
\item{level_1}{Top-level category.}
\item{level_2}{Second-level category.}
\item{level_3}{Child-level category.}
}
}
\source{
\href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-category-filtering.htm}{ArcGIS REST API category filtering}.
}
\description{
A dataset of categories that can be used to filter results from \code{\link[=arc_geo]{arc_geo()}},
\code{\link[=arc_geo_multi]{arc_geo_multi()}} and \code{\link[=arc_geo_categories]{arc_geo_categories()}}.
}
\details{
See \href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-category-filtering.htm}{ArcGIS REST API category filtering} for details and
examples.

The ArcGIS geocoding service supports searches for many types of addresses
and places around the world, so applications do not need to anticipate the
types of places that users may search for. However, ambiguous searches can
match many different places and produce unexpected results. For example, a
search for a city may match a street name, or an airport code may match a
country abbreviation.

In these cases, the \code{category} argument can filter out unwanted results. It
limits the types of places that the service searches for, which can eliminate
false-positive matches and speed up the search.

The dataset lists categories at three hierarchy levels (\code{level_1}, \code{level_2}
and \code{level_3}). If a \code{level_1} category is requested (for example, \code{POI}),
its child categories may also be included in the results.
}
\note{
Data extracted on \strong{15 January 2026}.
}
\examples{
\dontshow{if (arcgeocoder_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
# Get all possible values.
data("arc_categories")
arc_categories

# Use categories.

sea_1 <- arc_geo("sea",
  custom_query = list(outFields = c("LongLabel", "Type")),
  limit = 2
)

dplyr::glimpse(sea_1)

# Categories can disambiguate the result.

sea_2 <- arc_geo("sea",
  custom_query = list(outFields = c("LongLabel", "Type")),
  limit = 2, category = "Food"
)

dplyr::glimpse(sea_2)

# Use multiple categories.
sea_3 <- arc_geo("sea",
  custom_query = list(outFields = c("LongLabel", "Type")),
  sourcecountry = "UK", limit = 5,
  category = c("Amusement Park", "Aquarium")
)

dplyr::glimpse(sea_3)
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=arc_geo]{arc_geo()}}, \code{\link[=arc_geo_multi]{arc_geo_multi()}} and \code{\link[=arc_geo_categories]{arc_geo_categories()}}.
}
\keyword{datasets}
