% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-reverse-geo.R
\encoding{UTF-8}
\name{arc_reverse_geo}
\alias{arc_reverse_geo}
\title{Reverse geocode coordinates with the ArcGIS REST API}
\usage{
arc_reverse_geo(
  x,
  y,
  address = "address",
  full_results = FALSE,
  return_coords = TRUE,
  verbose = FALSE,
  progressbar = TRUE,
  outsr = NULL,
  langcode = NULL,
  featuretypes = NULL,
  locationtype = NULL,
  custom_query = list()
)
}
\arguments{
\item{x}{A numeric vector of longitude values in the range
\eqn{\left[-180, 180 \right]}.}

\item{y}{A numeric vector of latitude values in the range
\eqn{\left[-90, 90 \right]}.}

\item{address}{Name of the address column in the output. The default is
\code{"address"}.}

\item{full_results}{A logical value indicating whether to return all
available API fields. The default, \code{FALSE}, returns only latitude,
longitude and address.}

\item{return_coords}{A logical value indicating whether to return the input
coordinates with the results.}

\item{verbose}{A logical value indicating whether to display API request
details.}

\item{progressbar}{A logical value indicating whether to display a progress
bar for multiple queries.}

\item{outsr}{Spatial reference of the output coordinates. The default is
\code{NULL}, which uses the service default. See \strong{Details} and
\link{arc_spatial_references}.}

\item{langcode}{Language of the returned addresses.}

\item{featuretypes}{A character vector that limits the possible match types.
The default is \code{NULL}, which does not filter by feature type. See
\strong{Details}.}

\item{locationtype}{Location represented by the output geometry when
\code{featuretypes} is \code{"PointAddress"} or \code{"Subaddress"}. Valid values are
\code{"rooftop"} and \code{"street"}. The default is \code{NULL}.}

\item{custom_query}{A named list with additional API parameters.}
}
\value{
A \link[tibble:tbl_df]{tibble} with one match for each coordinate pair. The API
output fields \code{x} and \code{y} are named \code{lon} and \code{lat}. These coordinates
correspond to the matched feature and may differ from the input \code{x} and \code{y}
values.

See \href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-service-output.htm}{ArcGIS REST API output} for field details.
}
\description{
Converts longitude and latitude values into addresses. Latitudes must be in
the range \eqn{\left[-90, 90 \right]} and longitudes in the range
\eqn{\left[-180, 180 \right]}. Returns one match for each coordinate pair.
}
\details{
See the \href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-reverse-geocode.htm}{ArcGIS REST API documentation} for more
information and valid values.
}
\section{\code{outsr}}{
The spatial reference can be specified as a well-known ID (WKID). If not
specified, the spatial reference of the output locations is the same as that
of the service (WGS 84, that is, WKID 4326).

See \link{arc_spatial_references} for values and examples.
}

\section{\code{featuretypes}}{
See \code{vignette("feature-types", package = "arcgeocoder")} for a detailed
explanation of this argument.

This argument restricts the feature types returned by a reverse geocoding
request. Possible values are \code{"StreetInt"}, \code{"DistanceMarker"},
\code{"StreetAddress"}, \code{"StreetName"}, \code{"POI"}, \code{"Subaddress"},
\code{"PointAddress"}, \code{"Postal"} and \code{"Locality"}.

Supply multiple values as a character vector, for example,
\code{c("PointAddress", "StreetAddress")}.
}

\examples{
\dontshow{if (arcgeocoder_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
arc_reverse_geo(x = -73.98586, y = 40.75728)

# Several coordinate pairs.
arc_reverse_geo(x = c(-73.98586, -3.188375), y = c(40.75728, 55.95335))

# Use additional API options.
sev <- arc_reverse_geo(
  x = c(-73.98586, -3.188375),
  y = c(40.75728, 55.95335),
  # Restrict results to specific feature types.
  featuretypes = "POI,StreetInt",
  # Return results in this WKID.
  outsr = 102100,
  verbose = TRUE, full_results = TRUE
)

dplyr::glimpse(sev)
}
\dontshow{\}) # examplesIf}
}
\references{
\href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-reverse-geocode.htm}{ArcGIS REST API \code{reverseGeocode}}.
}
\seealso{
Geocoding and reverse geocoding functions:
\code{\link[=arc_geo]{arc_geo()}},
\code{\link[=arc_geo_categories]{arc_geo_categories()}},
\code{\link[=arc_geo_multi]{arc_geo_multi()}}
}
\concept{geocoders}
