% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{arc_spatial_references}
\alias{arc_spatial_references}
\title{Spatial references supported by the ArcGIS REST API}
\format{
A \link[tibble:tbl_df]{tibble} with
9,608 rows
and eight variables:
\describe{
\item{projtype}{Projection type (\code{"ProjectedCoordinateSystems"},
\code{"GeographicCoordinateSystems"} or \code{"VerticalCoordinateSystems"}).}
\item{wkid}{Well-known ID (WKID).}
\item{latestWkid}{Current WKID if \code{wkid} is deprecated.}
\item{authority}{WKID authority (Esri or EPSG).}
\item{deprecated}{Whether \code{wkid} is deprecated.}
\item{description}{Human-readable description of the spatial reference.}
\item{areaname}{Area of use of the spatial reference.}
\item{wkt}{Well-known text (WKT) representation of the spatial reference.
Useful when working with \CRANpkg{sf} or \CRANpkg{terra}.}
}
}
\source{
\href{https://github.com/Esri/projection-engine-db-doc}{Esri Projection Engine factory}.
}
\description{
A dataset of coordinate reference systems (CRSs) supported by the ArcGIS
REST API.
}
\details{
This dataset is useful when using the \code{outsr} argument.

Some projection IDs have changed over time. For example, Web Mercator
\code{wkid = 102100} is deprecated and its current equivalent is \code{wkid = 3857}.
Both values work and return equivalent results.
}
\note{
Data extracted on \strong{15 January 2026}.
}
\examples{
\dontshow{if (arcgeocoder_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
# Get all possible values.
data("arc_spatial_references")
arc_spatial_references

# Find the deprecated Web Mercator WKID.
library(dplyr)
wkid <- arc_spatial_references |>
  filter(latestWkid == 3857 & deprecated) |>
  slice(1)

glimpse(wkid)

add <- arc_geo("London, United Kingdom", outsr = wkid$wkid)

# Compare `lat`, `lon`, `wkid` and the current ID in `latestWkid`.
add |>
  select(lat, lon, wkid, latestWkid) |>
  glimpse()

# Look up the deprecated WKID.

try(sf::st_crs(wkid$wkid))

# Look up the current WKID.
try(sf::st_crs(wkid$latestWkid))

# Look up the WKT definition.
try(sf::st_crs(wkid$wkt))
}
\dontshow{\}) # examplesIf}
}
\seealso{
\itemize{
\item \code{\link[sf:st_crs]{sf::st_crs()}} inspects coordinate reference systems.
\item \code{\link[=arc_geo]{arc_geo()}}, \code{\link[=arc_geo_multi]{arc_geo_multi()}}, \code{\link[=arc_geo_categories]{arc_geo_categories()}} and
\code{\link[=arc_reverse_geo]{arc_reverse_geo()}} accept spatial references.
}
}
\keyword{datasets}
