% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgeocoder-package.R
\docType{package}
\encoding{UTF-8}
\name{arcgeocoder-package}
\alias{arcgeocoder}
\alias{arcgeocoder-package}
\title{arcgeocoder: Address and Coordinate Search with the 'ArcGIS' REST API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a lightweight interface to the 'ArcGIS' REST API for converting addresses and structured address components into geographic coordinates, finding places by category and converting coordinates into addresses. It uses the ArcGIS service documented at \url{https://developers.arcgis.com/rest/geocode/api-reference/overview-world-geocoding-service.htm}. No API key is required.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/arcgeocoder/}
  \item \url{https://github.com/dieghernan/arcgeocoder}
  \item Report bugs at \url{https://github.com/dieghernan/arcgeocoder/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]
}

}
\keyword{internal}
