% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_geo.R
\encoding{UTF-8}
\name{arc_geo}
\alias{arc_geo}
\title{Geocoding using the ArcGIS REST API}
\usage{
arc_geo(
  address,
  lat = "lat",
  long = "lon",
  limit = 1,
  full_results = FALSE,
  return_addresses = TRUE,
  verbose = FALSE,
  progressbar = TRUE,
  outsr = NULL,
  langcode = NULL,
  sourcecountry = NULL,
  category = NULL,
  custom_query = list()
)
}
\arguments{
\item{address}{Single-line address text (e.g.
\code{"1600 Pennsylvania Ave NW, Washington"}) or a vector of addresses
(e.g. \code{c("Madrid", "Barcelona")}).}

\item{lat}{Latitude column name in the output data (default \code{"lat"}).}

\item{long}{Longitude column name in the output data (default \code{"lon"}).}

\item{limit}{Maximum number of results to return per input address. Each
query has a hard API limit of 50 results.}

\item{full_results}{Logical. If \code{TRUE}, return all available API fields
via \verb{outFields=*}. Default is \code{FALSE}.}

\item{return_addresses}{Logical. If \code{TRUE}, keep input query in output.}

\item{verbose}{Logical. If \code{TRUE}, output process messages to the console.}

\item{progressbar}{Logical. If \code{TRUE}, show a progress bar for multiple
points.}

\item{outsr}{The spatial reference of the \code{x} and \code{y} coordinates returned
by a geocode request. By default, it is \code{NULL} (i.e. the argument will not
be used in the query). See \strong{Details} and \link{arc_spatial_references}.}

\item{langcode}{Sets the language in which reverse-geocoded addresses are
returned.}

\item{sourcecountry}{Country filter using ISO codes (e.g. \code{"USA"}).
Multiple values can be specified (comma-separated).}

\item{category}{Place or address type used as a filter. Multiple values are
accepted (e.g. \code{c("Cinema", "Museum")}). See \link{arc_categories}.}

\item{custom_query}{Additional API parameters as named list values.}
}
\value{
A \link[tibble:tbl_df]{tibble} object with the results. See output details in  \href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-service-output.htm}{ArcGIS REST API service output}.
}
\description{
Geocodes addresses given as character values. This function returns the
\link[tibble:tbl_df]{tibble} object associated with the query.

This function uses the \code{SingleLine} approach detailed in the
\href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-find-address-candidates.htm}{ArcGIS REST docs}. For multi-field queries (i.e.
using specific address components), use \code{\link[=arc_geo_multi]{arc_geo_multi()}}.
}
\details{
See the \href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-find-address-candidates.htm}{ArcGIS REST docs} for more information and
valid values.
}
\section{\code{outsr}}{
The spatial reference can be specified as a well-known ID (WKID). If not
specified, the spatial reference of the output locations is the same as that
of the service (WGS84, i.e. WKID = 4326).

See \link{arc_spatial_references} for values and examples.
}

\examples{
\dontshow{if (arcgeocoder_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
arc_geo("Madrid, Spain")

library(dplyr)

# Several addresses with additional output fields.
with_params <- arc_geo(c("Madrid", "Barcelona"),
  custom_query = list(outFields = c("LongLabel", "CntryName"))
)

with_params |>
  select(lat, lon, CntryName, LongLabel)

# With options: restrict search to the USA.
with_params_usa <- arc_geo(c("Madrid", "Barcelona"),
  sourcecountry = "USA",
  custom_query = list(outFields = c("LongLabel", "CntryName"))
)

with_params_usa |>
  select(lat, lon, CntryName, LongLabel)
}
\dontshow{\}) # examplesIf}
}
\references{
\href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-find-address-candidates.htm}{ArcGIS REST \code{findAddressCandidates}}.
}
\seealso{
\code{\link[tidygeocoder:geo]{tidygeocoder::geo()}}

Other functions for geocoding:
\code{\link[=arc_geo_categories]{arc_geo_categories()}},
\code{\link[=arc_geo_multi]{arc_geo_multi()}},
\code{\link[=arc_reverse_geo]{arc_reverse_geo()}}
}
\concept{geocoding}
