% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_geo_categories.R
\encoding{UTF-8}
\name{arc_geo_categories}
\alias{arc_geo_categories}
\title{Geocode places in a given area by category}
\usage{
arc_geo_categories(
  category,
  x = NULL,
  y = NULL,
  bbox = NULL,
  name = NULL,
  lat = "lat",
  long = "lon",
  limit = 1,
  full_results = FALSE,
  verbose = FALSE,
  custom_query = list(),
  ...
)
}
\arguments{
\item{category}{A place or address type that can be used to filter results.
Several values can also be supplied as a vector (i.e.
\code{c("Cinema", "Museum")}), which performs one call for each value. See
\strong{Details}.}

\item{x}{Longitude values in numeric format. Must be in the range
\eqn{\left[-180, 180 \right]}.}

\item{y}{Latitude values in numeric format. Must be in the range
\eqn{\left[-90, 90 \right]}.}

\item{bbox}{A numeric vector of longitude and latitude
\code{c(minX, minY, maxX, maxY)} that restricts the search area.
See \strong{Details}.}

\item{name}{Optionally, a string indicating the name or address of the
desired results.}

\item{lat}{Latitude column name in the output data (default \code{"lat"}).}

\item{long}{Longitude column name in the output data (default \code{"lon"}).}

\item{limit}{Maximum number of results per query. ArcGIS API limits a single
request to 50 results.}

\item{full_results}{Logical. If \code{TRUE}, return all available API fields
via \verb{outFields=*}. Default is \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE}, output process messages to the console.}

\item{custom_query}{Additional API parameters as named list values.}

\item{...}{
  Arguments passed on to \code{\link{arc_geo}}
  \describe{
    \item{\code{sourcecountry}}{Country filter using ISO codes (e.g. \code{"USA"}).
Multiple values can be specified (comma-separated).}
    \item{\code{outsr}}{The spatial reference of the \code{x} and \code{y} coordinates returned
by a geocode request. By default, it is \code{NULL} (i.e. the argument will not
be used in the query). See \strong{Details} and \link{arc_spatial_references}.}
    \item{\code{langcode}}{Sets the language in which reverse-geocoded addresses are
returned.}
  }}
}
\value{
A \link[tibble:tbl_df]{tibble} object with the results. See output details in  \href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-service-output.htm}{ArcGIS REST API service output}.
}
\description{
This function is useful for extracting places with a given category (or list
of categories) near or within a given location or area. This is a wrapper
of \code{\link[=arc_geo]{arc_geo()}}, but it is vectorized over \code{category}.

See \link{arc_categories} for a detailed explanation and available values.

\strong{Note:} to obtain results, provide one of the following:
\itemize{
\item A pair of coordinates (\code{x} and \code{y} arguments) used as a reference for
geocoding.
\item A bounding box via the \code{bbox} argument defining a desired extent for
results.
}

It is possible to combine both approaches (i.e. providing \code{x}, \code{y} and
\code{bbox} values) to improve the geocoding process. See \strong{Examples}.
}
\details{
Bounding boxes can be located using online tools, such as
\href{https://boundingbox.klokantech.com/}{Bounding Box Tool}.

For a full list of valid categories see \link{arc_categories}. This function is
vectorized over \code{category}, which means it performs one independent call
to \code{\link[=arc_geo]{arc_geo()}} for each \code{category} value.

\code{arc_geo_categories()} also understands a single string of categories
separated by commas (\code{"Cinema,Museum"}), which is treated internally as
\code{c("Cinema", "Museum")}.
}
\section{\code{outsr}}{
The spatial reference can be specified as a well-known ID (WKID). If not
specified, the spatial reference of the output locations is the same as that
of the service (WGS84, i.e. WKID = 4326).

See \link{arc_spatial_references} for values and examples.
}

\examples{
\dontshow{if (arcgeocoder_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
# Full workflow: gas stations near Carabanchel, Madrid.

# Get Carabanchel.
carab <- arc_geo("Carabanchel, Madrid, Spain")

# CRS.
carab_crs <- unique(carab$latestWkid)

library(ggplot2)

base_map <- ggplot(carab) +
  geom_point(aes(lon, lat), size = 5, color = "red") +
  geom_rect(aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax),
    fill = NA,
    color = "blue"
  ) +
  coord_sf(crs = carab_crs)

# Example 1: Search near Carabanchel (not restricted).
ex1 <- arc_geo_categories("Gas Station",
  # Location.
  x = carab$lon, y = carab$lat,
  limit = 50, full_results = TRUE
)

# Reduce labels to the most common ones.
library(dplyr)

labs <- ex1 |>
  count(ShortLabel) |>
  slice_max(n = 7, order_by = n) |>
  pull(ShortLabel)

base_map +
  geom_point(data = ex1, aes(lon, lat, color = ShortLabel)) +
  scale_color_discrete(breaks = labs) +
  labs(
    title = "Example 1",
    subtitle = "Search near (points may be far away)"
  )

# Example 2: Include part of the name for different results.
ex2 <- arc_geo_categories("Gas Station",
  # Name.
  name = "Repsol",
  # Location.
  x = carab$lon, y = carab$lat,
  limit = 50, full_results = TRUE
)

base_map +
  geom_point(data = ex2, aes(lon, lat, color = ShortLabel)) +
  labs(
    title = "Example 2",
    subtitle = "Search near with name"
  )

# Example 3: Search within a bounding box.
ex3 <- arc_geo_categories("Gas Station",
  name = "Repsol",
  bbox = c(carab$xmin, carab$ymin, carab$xmax, carab$ymax),
  limit = 50, full_results = TRUE
)

base_map +
  geom_point(data = ex3, aes(lon, lat, color = ShortLabel)) +
  labs(
    title = "Example 3",
    subtitle = "Search near with name and bounding box"
  )
}
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-category-filtering.htm}{ArcGIS REST Category filtering}.

\link{arc_categories}

Other functions for geocoding:
\code{\link[=arc_geo]{arc_geo()}},
\code{\link[=arc_geo_multi]{arc_geo_multi()}},
\code{\link[=arc_reverse_geo]{arc_reverse_geo()}}
}
\concept{geocoding}
