% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_reverse_geo.R
\encoding{UTF-8}
\name{arc_reverse_geo}
\alias{arc_reverse_geo}
\title{Reverse geocoding using the ArcGIS REST API}
\usage{
arc_reverse_geo(
  x,
  y,
  address = "address",
  full_results = FALSE,
  return_coords = TRUE,
  verbose = FALSE,
  progressbar = TRUE,
  outsr = NULL,
  langcode = NULL,
  featuretypes = NULL,
  locationtype = NULL,
  custom_query = list()
)
}
\arguments{
\item{x}{Longitude values in numeric format. Must be in the range
\eqn{\left[-180, 180 \right]}.}

\item{y}{Latitude values in numeric format. Must be in the range
\eqn{\left[-90, 90 \right]}.}

\item{address}{Output address column name (default \code{"address"}).}

\item{full_results}{Logical. If \code{TRUE}, return all available API fields.
\code{FALSE} (default) returns latitude, longitude and address only.}

\item{return_coords}{Logical. If \code{TRUE}, return input coordinates with
results.}

\item{verbose}{Logical. If \code{TRUE}, output process messages to the console.}

\item{progressbar}{Logical. If \code{TRUE}, show a progress bar for multiple
points.}

\item{outsr}{The spatial reference of the \code{x} and \code{y} coordinates returned
by a geocode request. By default, it is \code{NULL} (i.e. the argument will not
be used in the query). See \strong{Details} and \link{arc_spatial_references}.}

\item{langcode}{Sets the language in which reverse-geocoded addresses are
returned.}

\item{featuretypes}{This argument limits the possible match types returned.
By default, it is \code{NULL} (i.e. the argument will not be used in the query).
See \strong{Details}.}

\item{locationtype}{Specifies whether the output geometry of
\code{featuretypes = "PointAddress"} or \code{featuretypes = "Subaddress"} matches
should be the rooftop point or street entrance location. Valid values are
\code{NULL} (i.e. not using the argument in the query), \code{rooftop} and \code{street}.}

\item{custom_query}{API-specific arguments to be used, passed as a named
list.}
}
\value{
A \link[tibble:tbl_df]{tibble} with the corresponding results. The \code{x} and \code{y}
values returned by the API are named \code{lon} and \code{lat}. Note that these
coordinates correspond to the geocoded feature and may differ from the \code{x}
and \code{y} values provided as inputs.

See the details of the output in
\href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-service-output.htm}{ArcGIS REST API service output}.
}
\description{
Generates an address from a latitude and longitude. Latitudes must be in the
range \eqn{\left[-90, 90 \right]} and longitudes in the range
\eqn{\left[-180, 180 \right]}. This function returns the
\link[tibble:tbl_df]{tibble} associated with the query.
}
\details{
See the \href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-reverse-geocode.htm}{ArcGIS REST docs} for more information and
valid values.
}
\section{\code{outsr}}{
The spatial reference can be specified as a well-known ID (WKID). If not
specified, the spatial reference of the output locations is the same as that
of the service (WGS84, i.e. WKID = 4326).

See \link{arc_spatial_references} for values and examples.
}

\section{\code{featuretypes}}{
See \code{vignette("featuretypes", package = "arcgeocoder")} for a detailed
explanation of this argument.

This argument may be used to filter the type of feature returned when
geocoding. Possible values are:
\itemize{
\item \code{"StreetInt"}
\item \code{"DistanceMarker"}
\item \code{"StreetAddress"}
\item \code{"StreetName"}
\item \code{"POI"}
\item \code{"Subaddress"}
\item \code{"PointAddress"}
\item \code{"Postal"}
\item \code{"Locality"}
}

It is also possible to use several values as a vector
(\code{featuretypes = c("PointAddress", "StreetAddress")}).
}

\examples{
\dontshow{if (arcgeocoder_check_access()) withAutoprint(\{ # examplesIf}
\donttest{

arc_reverse_geo(x = -73.98586, y = 40.75728)

# Several coordinates.
arc_reverse_geo(x = c(-73.98586, -3.188375), y = c(40.75728, 55.95335))

# With options: use additional arguments.
sev <- arc_reverse_geo(
  x = c(-73.98586, -3.188375),
  y = c(40.75728, 55.95335),
  # Restrict to these features.
  featuretypes = "POI,StreetInt",
  # Return results in this WKID.
  outsr = 102100,
  verbose = TRUE, full_results = TRUE
)

dplyr::glimpse(sev)
}
\dontshow{\}) # examplesIf}
}
\references{
\href{https://developers.arcgis.com/rest/geocode/api-reference/geocoding-reverse-geocode.htm}{ArcGIS REST \code{reverseGeocode}}.
}
\seealso{
\code{\link[tidygeocoder:reverse_geo]{tidygeocoder::reverse_geo()}}

Other functions for geocoding:
\code{\link[=arc_geo]{arc_geo()}},
\code{\link[=arc_geo_categories]{arc_geo_categories()}},
\code{\link[=arc_geo_multi]{arc_geo_multi()}}
}
\concept{geocoding}
