% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{arc_spatial_references}
\alias{arc_spatial_references}
\title{Esri (ArcGIS) spatial reference database}
\format{
A \link[tibble:tbl_df]{tibble} with
9,608 rows
and fields:
\describe{
\item{projtype}{Projection type (\code{"ProjectedCoordinateSystems"},
\code{"GeographicCoordinateSystems"} or \code{"VerticalCoordinateSystems"}).}
\item{wkid}{Well-Known ID.}
\item{latestWkid}{Most recent \code{wkid}, if \code{wkid} is deprecated.}
\item{authority}{\code{wkid} authority (Esri or EPSG).}
\item{deprecated}{Logical indicating whether \code{wkid} is deprecated.}
\item{description}{Human-readable description of the \code{wkid}.}
\item{areaname}{Use area of the \code{wkid}.}
\item{wkt}{Representation of \code{wkid} in Well-Known Text (WKT). Useful when
working with \CRANpkg{sf} or \CRANpkg{terra}.}
}
}
\source{
\href{https://github.com/Esri/projection-engine-db-doc}{Esri Projection Engine factory}
}
\description{
Database of available spatial references (CRS) in \link[tibble:tbl_df]{tibble}
format.
}
\details{
This database is useful when using the \code{outsr} argument.

Some projection IDs have changed over time. For example, Web Mercator
\code{wkid = 102100} is deprecated and is currently \code{wkid = 3857}. However, both
values work and return similar results.
}
\note{
Data extracted on \strong{15 January 2026}.
}
\examples{
\dontshow{if (arcgeocoder_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
# Get all possible values.
data("arc_spatial_references")
arc_spatial_references

# Request with deprecated Web Mercator.
library(dplyr)
wkid <- arc_spatial_references |>
  filter(latestWkid == 3857 & deprecated) |>
  slice(1)

glimpse(wkid)

add <- arc_geo("London, United Kingdom", outsr = wkid$wkid)

# Note values for `lat`, `lon` and `wkid`. `latestWkid` gives the current
# valid `wkid`.
add |>
  select(lat, lon, wkid, latestWkid) |>
  glimpse()

# Try with `sf`.

try(sf::st_crs(wkid$wkid))

# Try the latest WKID.
try(sf::st_crs(wkid$latestWkid))

# Or use WKT.
try(sf::st_crs(wkid$wkt))
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[sf:st_crs]{sf::st_crs()}}

Other datasets:
\code{\link{arc_categories}}
}
\concept{datasets}
