% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgeocoder-package.R
\docType{package}
\name{arcgeocoder-package}
\alias{arcgeocoder}
\alias{arcgeocoder-package}
\title{arcgeocoder: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dieghernan/arcgeocoder}
  \item Report bugs at \url{https://github.com/dieghernan/arcgeocoder/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

}
\keyword{internal}
