% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgeocoder-package.R
\docType{package}
\name{arcgeocoder-package}
\alias{arcgeocoder}
\alias{arcgeocoder-package}
\title{arcgeocoder: Geocoding with the 'ArcGIS' REST API Service}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Lite interface for geocoding and reverse geocoding with the 'ArcGIS' REST API service \url{https://developers.arcgis.com/rest/geocode/api-reference/overview-world-geocoding-service.htm}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/arcgeocoder/}
  \item \url{https://github.com/dieghernan/arcgeocoder}
  \item Report bugs at \url{https://github.com/dieghernan/arcgeocoder/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

}
\keyword{internal}
