% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgeocoder-check-access.R
\encoding{UTF-8}
\name{arcgeocoder_check_access}
\alias{arcgeocoder_check_access}
\title{Check access to the ArcGIS REST API}
\usage{
arcgeocoder_check_access()
}
\value{
\code{TRUE} if the service is accessible, otherwise \code{FALSE}.
}
\description{
Checks whether the current \R session can access the ArcGIS REST API at
\url{https://developers.arcgis.com/rest/geocode/api-reference/overview-world-geocoding-service.htm}.
}
\examples{
\donttest{
arcgeocoder_check_access()
}
}
\keyword{internal}
