% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgeocoder_check_access.R
\name{arcgeocoder_check_access}
\alias{arcgeocoder_check_access}
\title{Check access to ArcGIS REST}
\usage{
arcgeocoder_check_access()
}
\value{
a logical.
}
\description{
Check if \strong{R} has access to resources at ArcGIS REST API
\url{https://developers.arcgis.com/rest/geocode/api-reference/overview-world-geocoding-service.htm}.
}
\examples{
\donttest{
arcgeocoder_check_access()
}
}
\concept{api_management}
\keyword{internal}
